bug fix
 - firing transition has other color
 - there is no pin available
 - threshold fields change to Integer
 - check capacity stuffs
 - english corrections

+++ src/main/java/org/pneditor/petrinet/xml/XmlArduinoComponentSettings.java
@@ -12,10 +12,10 @@ public class XmlArduinoComponentSettings {
 
     // ANALOG
     @XmlElement(name = "bottomThreshold")
-    public Double bottomThreshold;
+    public Integer bottomThreshold;
 
     @XmlElement(name = "upThreshold")
-    public Double upThreshold;
+    public Integer upThreshold;
 
     // MESSAGE + CUSTOM SYSEX
     @XmlElement(name = "message")
